requirements.txt
click==8.0.4
Flask==2.0.3
greenlet==1.1.2
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
SQLAlchemy==1.4.32
Werkzeug==2.0.3
